feat: implement deposit to a connext address
