+++ src/common/currencyUtil.ts
@@ -44,3 +44,6 @@ export const getCurrencyFullName = (currency: string): string => {
   const fullName = CURRENCY_DICTIONARY[currency.toUpperCase()]?.fullName;
   return fullName || currency;
 };
+
+export const isLnd = (currency: string): boolean =>
+  ["BTC", "LTC"].includes(currency);
